backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart